backend(qa): reorganize code
