+++ apps/backend-functions/src/internals/index.ts
@@ -1,0 +1,3 @@
+/**
+ * Define all the internal tools and libraries we use in the backend.
+ */
