+++ apps/backend-functions/src/internals/pdf.ts
@@ -5,8 +5,8 @@
  * and in a firebase function.
  */
 import { groupBy, sortBy, isEmpty } from 'lodash';
-import { asIDMap, Delivery, IDMap, Material, Organization, Stakeholder, Step } from './data/types';
-import { getCollection, getDocument } from './data/internals';
+import { asIDMap, Delivery, IDMap, Material, Organization, Stakeholder, Step } from '../data/types';
+import { getCollection, getDocument } from '../data/internals';
 
 const PdfPrinter = require('pdfmake');
